Sema: Move some code to CSDisjunction.cpp
The argument label-based pruning is similar to the type-based pruning
that I added, and I'm going to try to consolidate the duplication
a bit.
