+++ lib/Sema/CSSolver.cpp
@@ -1167,109 +1167,6 @@ void ConstraintSystem::collectDisjunctions(
   }
 }
 
-ConstraintSystem::SolutionKind
-ConstraintSystem::filterDisjunction(
-    Constraint *disjunction, bool restoreOnFail,
-    llvm::function_ref<bool(Constraint *)> pred) {
-  assert(disjunction->getKind() == ConstraintKind::Disjunction);
-
-  SmallVector<Constraint *, 4> constraintsToRestoreOnFail;
-  unsigned choiceIdx = 0;
-  unsigned numEnabledTerms = 0;
-  ASTContext &ctx = getASTContext();
-  for (unsigned constraintIdx : indices(disjunction->getNestedConstraints())) {
-    auto constraint = disjunction->getNestedConstraints()[constraintIdx];
-
-    // Skip already-disabled constraints. Let's treat disabled
-    // choices which have a fix as "enabled" ones here, so we can
-    // potentially infer some type information from them.
-    if (constraint->isDisabled() && !constraint->getFix())
-      continue;
-
-    if (pred(constraint)) {
-      ++numEnabledTerms;
-      choiceIdx = constraintIdx;
-      continue;
-    }
-
-    if (isDebugMode()) {
-      auto indent = (solverState ? solverState->getCurrentIndent() : 0) + 4;
-      llvm::errs().indent(indent) << "(disabled disjunction term ";
-      constraint->print(llvm::errs(), &ctx.SourceMgr, indent);
-      llvm::errs().indent(indent) << ")\n";
-    }
-
-    if (!constraint->isDisabled()) {
-      if (restoreOnFail)
-        constraintsToRestoreOnFail.push_back(constraint);
-      else if (solverState)
-        solverState->disableConstraint(constraint);
-      else
-        constraint->setDisabled();
-    }
-  }
-
-  if (numEnabledTerms == 0)
-    return SolutionKind::Error;
-
-  if (restoreOnFail) {
-    for (auto constraint : constraintsToRestoreOnFail) {
-      if (solverState)
-        solverState->disableConstraint(constraint);
-      else
-        constraint->setDisabled();
-    }
-  }
-
-  if (numEnabledTerms == 1) {
-    // Only a single constraint remains. Retire the disjunction and make
-    // the remaining constraint active.
-    auto choice = disjunction->getNestedConstraints()[choiceIdx];
-
-    // This can only happen when subscript syntax is used to lookup
-    // something which doesn't exist in type marked with
-    // `@dynamicMemberLookup`.
-    // Since filtering currently runs as part of the `applicable function`
-    // constraint processing, "keypath dynamic member lookup" choice can't
-    // be attempted in-place because that would also try to operate on that
-    // constraint, so instead let's keep the disjunction, but disable all
-    // unviable choices.
-    if (choice->getOverloadChoice().isKeyPathDynamicMemberLookup()) {
-      for (auto *currentChoice : disjunction->getNestedConstraints()) {
-        if (currentChoice->isDisabled())
-          continue;
-
-        if (currentChoice != choice)
-          solverState->disableConstraint(currentChoice);
-      }
-      return SolutionKind::Solved;
-    }
-
-    // Retire the disjunction. It's been solved.
-    retireConstraint(disjunction);
-
-    // Note the choice we made and simplify it. This introduces the
-    // new constraint into the system.
-    if (disjunction->shouldRememberChoice()) {
-      recordDisjunctionChoice(disjunction->getLocator(), choiceIdx);
-    }
-
-    if (isDebugMode()) {
-      auto indent = (solverState ? solverState->getCurrentIndent() : 0) + 4;
-      llvm::errs().indent(indent)
-          << "(introducing single enabled disjunction term ";
-      choice->print(llvm::errs(), &ctx.SourceMgr, indent);
-      llvm::errs().indent(indent) << ")\n";
-    }
-
-    simplifyDisjunctionChoice(choice);
-
-    return failedConstraint ? SolutionKind::Unsolved : SolutionKind::Solved;
-  }
-
-  return SolutionKind::Unsolved;
-}
-
 std::optional<std::pair<Constraint *, unsigned>>
 ConstraintSystem::findConstraintThroughOptionals(
     TypeVariableType *typeVar, OptionalWrappingDirection optionalDirection,
